feat: mutation with auth

+++ src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
+import getUserId from "../utils/getUserId";
 
 const Mutation = {
   login: async (parent, { data }, { prisma }, info) => {
@@ -47,22 +48,21 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, "supercoolsecret")
     }
   },
-  deleteUser: async (parent, { id }, { prisma }, info) => {
-    const userExists = await prisma.exists.User({ id });
-    if (!userExists) {
-      throw new Error("User not found.");
-    }
+  deleteUser: async (parent, { id }, { prisma, request }, info) => {
+    const userId = getUserId(request);
     return prisma.mutation.deleteUser({
-      where: { id }
+      where: { id: userId }
     }, info);
   },
-  updateUser: async (parent, { id, data }, { prisma }, info) => {
+  updateUser: async (parent, { id, data }, { prisma, request }, info) => {
+    const userId = getUserId(request);
     return prisma.mutation.updateUser({
-      where: { id },
+      where: { id: userId },
       data
     }, info);
   },
-  createPost: async (parent, { data }, { prisma }, info) => {
+  createPost: async (parent, { data }, { prisma, request }, info) => {
+    const userId = getUserId(request);
     const userExists = await prisma.exists.User({ id: data.author });
     if (!userExists) {
       throw new Error("User not found.");
@@ -74,30 +74,57 @@ const Mutation = {
         published: data.published,
         author: {
           connect: {
-            id: data.author
+            id: userId
           }
         }
       }
     }, info);
   },
-  deletePost(parent, { id }, { prisma }, info) {
+  deletePost: async (parent, { id }, { prisma, request }, info) => {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) {
+      throw new Error("Unable to delete post.");
+    }
+
     return prisma.mutation.deletePost({
-      where: { id }
+      where: {
+        id
+      }
     }, info);
   },
-  updatePost(parent, { id, data }, { prisma }, info) {
+  updatePost: async (parent, { id, data }, { prisma, request }, info) => {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) {
+      throw new Error("Unable to update post.");
+    }
+
     return prisma.mutation.updatePost({
       where: { id },
       data
     }, info);
   },
-  createComment(parent, { data }, { prisma }, info) {
+  createComment(parent, { data }, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createComment({
       data: {
         text: data.text,
         author: {
           connect: {
-            id: data.author
+            id: userId
           }
         },
         post: {
@@ -108,12 +135,36 @@ const Mutation = {
       }
     }, info);
   },
-  deleteComment(parent, { id }, { prisma }, info) {
+  deleteComment: async (parent, { id }, { prisma, request }, info) => {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!commentExists) {
+      throw new Error("Unable to delete comment.");
+    }
+
     return prisma.mutation.deleteComment({
       where: { id }
     }, info);
   },
-  updateComment(parent, { id, data }, { prisma }, info) {
+  updateComment: async (parent, { id, data }, { prisma, request }, info) => {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!commentExists) {
+      throw new Error("Unable to update comment.");
+    }
+
     return prisma.mutation.updateComment({
       where: { id },
       data
